Override RowAwarePrimaryKeyFactory with new components specific implementation
