+++ src/java/org/apache/cassandra/index/sai/disk/v9/V9OnDiskFormat.java
@@ -70,14 +70,14 @@ public class V9OnDiskFormat extends V8OnDiskFormat
     @Override
     public PrimaryKey.Factory newPrimaryKeyFactory(ClusteringComparator comparator)
     {
-        return new RowAwarePrimaryKeyFactory(comparator);
+        return new OptimizedRowAwarePrimaryKeyFactory(comparator);
     }
 
     @Override
     public PrimaryKeyMap.Factory newPrimaryKeyMapFactory(IndexComponents.ForRead perSSTableComponents, PrimaryKey.Factory primaryKeyFactory, SSTableReader sstable)
     {
-        assert primaryKeyFactory instanceof RowAwarePrimaryKeyFactory;
-        RowAwarePrimaryKeyFactory rowAwareFactory = (RowAwarePrimaryKeyFactory) primaryKeyFactory;
+        assert primaryKeyFactory instanceof OptimizedRowAwarePrimaryKeyFactory;
+        OptimizedRowAwarePrimaryKeyFactory rowAwareFactory = (OptimizedRowAwarePrimaryKeyFactory) primaryKeyFactory;
         return rowAwareFactory.hasClustering ? new WidePrimaryKeyMap.Factory(perSSTableComponents, rowAwareFactory, sstable)
                                              : new SkinnyPrimaryKeyMap.Factory(perSSTableComponents, rowAwareFactory, sstable);
     }
